refactor(kv-router): move potential load DTO to protocols
